requirements.txt
annotated-types==0.7.0
anyio==4.8.0
bcrypt==4.1.3
certifi==2025.1.31
click==8.1.8
colorama==0.4.6
coverage==7.6.12
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.111.0
fastapi-cli==0.0.7
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
Jinja2==3.1.5
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
orjson==3.10.15
packaging==24.2
passlib==1.7.4
pluggy==1.5.0
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.19.1
PyJWT==2.4.0
pytest==8.3.5
pytest-cov==6.0.0
pytest-mock==3.14.0
python-dotenv==1.0.1
python-multipart==0.0.20
PyYAML==6.0.2
rich==13.9.4
rich-toolkit==0.13.2
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.31
sqlmodel==0.0.19
starlette==0.37.2
typer==0.15.2
typing_extensions==4.12.2
ujson==5.10.0
uvicorn==0.30.1
watchfiles==1.0.4
websockets==15.0
